Add net::SdchSourceStream and net::SdchPolicyDelegate
This CL adds a SdchSourceStream which implements FilterSourceStream to do
sdch decoding. This CL additionally adds a SdchPolicyDelegate which will be
the glue for URLRequestHttpJob and SdchSourceStream.

This is a part of the effort to convert net::Filter into a pull-based
interface. See the linked bug for more details.

BUG=474859

Review-Url: https://chromiumcodereview.appspot.com/2368433002
Cr-Original-Commit-Position: refs/heads/master@{#426803}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab5a1f3c40b48b2a0498e6469f9f8d676a728994