Resolve a TSAN race in CertVerifyProc
TSAN gets unhappy because MultiThreadedCertVerifier abandons its
worker threads (because they may be blocked on OS APIs) when
shutting down tests, which interferes with ScopedFeatureList
shutdown code in //net.

This simply moves the check from being on the worker thread
when it's running, to being on the main thread when the
CertVerifyProc is constructed.

BUG=663850

Review-Url: https://codereview.chromium.org/2493593003
Cr-Original-Commit-Position: refs/heads/master@{#431164}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d3cd28a62874bbdc27be9134dd6fe53f6b8cd10