NQE: Notify observer as soon as it is added
When an EffectiveConnectionTypeObserver or an
RTTAndThroughputEstimatesObserver is added to NetworkQualityEstimator
(NQE), it is notified of the current effective connection type or
current RTT/throughput estimates.

Doing this makes the API simpler (the caller does not have to
call Get*() functions when registering as an observer). It also
reduces the need of exposing both Get*() and Add*() APIs to different
layers of Chromium.

BUG=638304

Review-Url: https://codereview.chromium.org/2491703003
Cr-Original-Commit-Position: refs/heads/master@{#432637}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 286d35bb3f13b27ada977c535d35fbe3977b9426