Add a URLRequest FTP fuzzer.
Also add a new method to FuzzedDataProvider to provide a random length
string, with a max length. Generating the string this way allows for
the fuzzer to mutate it more easily than the previous pattern of
picking length from the end of input and then choosing that many
characters from the start of the input.

FuzzedSockets are switched to use the new method.  It will cause them
to need to rebuild their corpuses, but should result in faster
convergence, longer term.

BUG=660869
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2469813002
Cr-Original-Commit-Position: refs/heads/master@{#433652}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7da0714bb1e387b30298047a9fa81c0ed2a2d13