Make DnsSocketPool copy RandIntCallback
DnsSocketPool is created by DnsClient, and passed a reference to the
RandIntCallback in DnsClient. It is then stored in DnsSession, which is
ref-counted (so running transactions can continue while DnsSession is
recreated on config changes).

Unfortunately, if the DnsClient itself is destroyed, the DnsSession
hangs around, holding on to the DnsSocketPool, which is holding on to a
reference to the RandIntCallback in the destroyed DnsClient.

Fix this by copying the RandIntCallback into the DnsSocketPool when it
is constructed.

(Also, rearrange destructor to satisfy the linter.)

BUG=664457

Review-Url: https://codereview.chromium.org/2533753003
Cr-Original-Commit-Position: refs/heads/master@{#434752}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2dd66098349196a1048403810c0faadf05072387