netlog: Don't enable recording if the "netlog" tracing category is disabled.
NetLog enables recording whenever an observer is
registered. TraceNetLogObserver would always register an observer, even
if the "netlog" tracing category is disabled. This is wasteful, since in
this case the trace events are dropped by TraceLog.

This CL only registers the observer when the category is enabled.

BUG=670732

Review-Url: https://codereview.chromium.org/2554483002
Cr-Original-Commit-Position: refs/heads/master@{#436368}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be0aa5179a641f9b087b119177be0ea26472358f