Update CertVerifyProcTest.LargeKey for Sierra (Mac OS 10.12)
On 10.12 a large key causes a recoverable error instead of a fatal error,
and one of the chain status codes is CSSMERR_TP_NOT_TRUSTED, so the
cert_status is now (CERT_STATUS_AUTHORITY_INVALID | CERT_STATUS_INVALID).
Update the test to just check if CERT_STATUS_INVALID is present in
cert_status so that it will pass on any version of mac.

BUG=629712

Review-Url: https://codereview.chromium.org/2550333003
Cr-Original-Commit-Position: refs/heads/master@{#436714}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99c0d7490554f43ed1b14d315a39be5f4c5e3fc0