[CacheTool] Add a "list_dups" command
Adds a new command to cachetool called "list_dups". It md5sums the response
bodies in the cache and prints out any duplicates along with their mime-type
and number of bytes. This is useful for helping to explore the possible
benefits to content-addressable caching.

BUG=665548

Review-Url: https://codereview.chromium.org/2421583002
Cr-Original-Commit-Position: refs/heads/master@{#438494}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ffc18a0d91f8d1f47f226cb0b837108ca6ea2cfa