Move ev_root_ca_metadata to read-only data segment
Due to a quirk in VC++ it is easy to accidentally prevent a const global
array from being placed in the read-only data segment. This change
removes a should-be-harmless 'const' keyword to work around this quirk
and move ~6,800 bytes to the .rdata (read-only) data segment.

VC++ bug is filed here:
https://connect.microsoft.com/VisualStudio/feedback/details/3117602

Other instances of this quirk are also being worked around.

BUG=677351

Review-Url: https://codereview.chromium.org/2607183002
Cr-Original-Commit-Position: refs/heads/master@{#440977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60cd9d9826b9611587c218b5bfb59f5a75767bb2