[Android WebView] Remove postOnPageFinished from shouldInterceptRequest.
We previously added a call to onPageFinished from shouldInterceptRequest
in https://codereview.chromium.org/1364863002/.
This was to avoid failing the test
AwContentsClientShouldInterceptRequestTest#testOnReceivedErrorCallback.
However, the cause of the test failing was that with that CL we no
longer post onPageFinished after receiving a response with an empty data
segment (because WebView no longer received
WebContentsObserver.didFinishLoad() callbacks for loads that resulted in
an empty response).
Since then, onPageFinished is again being posted after receiving an
empty response - this was fixed in
https://codereview.chromium.org/2563423004/

The current CL simply removes the unnecessary (and thus incorrect)
onPageFinished call from shouldInterceptRequest and adds a couple of
tests to ensure the original problem with onPageFinished doesn't occur
again.

See b/32629339 for additional discussions.

BUG=656919

Review-Url: https://codereview.chromium.org/2624443002
Cr-Original-Commit-Position: refs/heads/master@{#442621}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7363ea0b81a3a51eb1513761ecfee892f667fa64