net: fix multiple use-after-scope issues in net_unittests.
Most of the issues are due to the fact that Transaction destructor
accesses the request reference that is stack allocated and has
a shorter lifespan, which causes an invalid access to a potentially
reused memory.

The bug was found by AddressSanitizer with use-after-free check
enabled. It's currently being rolled out into Chrome, and this CL
is a part of a larger cleanup of existing failures.

BUG=649897

Review-Url: https://codereview.chromium.org/2635533002
Cr-Original-Commit-Position: refs/heads/master@{#443685}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0bfeb6b121c5cfd7658c485dcf2077826bad2f2b