Audit CT logs by requesting inclusion proofs for observed Signed Certificate Timestamps from the CT log that issued them.
To verify that a CT log behaves correctly and indeed
publishes all the certificates it committed to publishing,
it is necessary to verify that each observed log entry
(as denoted by an SCT and the corresponding
certificate) is included in the log.

Check for inclusion of observed SCTs by finding out
the leaf index of each observed SCT and requesting
an inclusion proof from the CT log for it over DNS
(using the LogDnsClient).

Note that no action is taken based on the inclusion
check results in this change. Follow-up changes
will add telemetry.

BUG=613495

Review-Url: https://codereview.chromium.org/2017563002
Cr-Original-Commit-Position: refs/heads/master@{#445513}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b5a833bca72a9f89e8c6e8c592037540eaf1411