Avoid unneeded refcount bump on disk_cache::File
A PostTaskAndReply in disk_cache::File implementation has unneeded
refcount bump, which is currently safe and going to be racy. This CL
removes the count bump to avoid the race on the upcoming change.

base::Callback retains the reference to disk_cache::File, and releases it
on the callback destruction. PostTaskAndReply currently guarantees that it
drops the reference to both `Task` callback and `Reply` callback on the origin
thread. However after a proposed change for PostTaskAndReply, the `Task`
callback will always be destroyed on the target thread of PostTaskAndReply.
That implies that the refcount of disk_cache::File will be dropped on a
worker thread, and causes a race on the refcount bump.
https://codereview.chromium.org/2657603004/

Review-Url: https://codereview.chromium.org/2657003003
Cr-Original-Commit-Position: refs/heads/master@{#446770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc974fef99b96b1a153fa7e7c35949de285c7e67