Improve HttpStreamFactory NetLog events
This CL introduces a SourceType for JobController to better account
for orphaned and preconnect jobs:

(1) For orphaned job, it will be correctly closed when the job
completes.
(2) For preconnect job, it will now appear in net-internals because
we take NetLog from HttpNetworkSession and log events there. Before
this CL, because preconnects do not have an associated stream request,
their events are not logged and it is hard to see whether a socket is
created due to preconnects.

BUG=679968
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2621983004
Cr-Original-Commit-Position: refs/heads/master@{#446985}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9255195315efa671012f0bd0da1afe685d70b2f6