Rename error enums and methods in SpdyFramer.
Motivation: for clarity, and to avoid confusion with ErrorCodeToString()
that I am planning to introduce.

This CL performs the following renames in SpdyFramer:
* enum SpdyError => enum SpdyFramerError;
* ErrorCodeToString() => SpdyFramerErrorToString().

This CL performs the following renames in SpdyFramer and in
Http2DecoderAdapter:
* error_code_ => spdy_framer_error_;
* error_code() => spdy_framer_error();

This CL lands server changes 144382359 and 145992550 by bnc.

BUG=488484

Review-Url: https://codereview.chromium.org/2665443003
Cr-Original-Commit-Position: refs/heads/master@{#447252}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04e0f45349341e37603aa31adfbbea9e81075129