Refactor two HTTP/2 stringify methods.
Move FrameTypeToString() and StatusCodeToString() from SpdyFramer to
spdy_protocol, because that is closer to the definition of enum FrameType, and
there is nothing specific to SpdyFramer about this method.

Rename StatusCodeToString() to ErrorCodeToString(), because SpdyGoAwayStatus and
SpdyRstStreamStatus enum values are referred to as "error codes" by the
specification.  A future CL will actually rename these enums to SpdyErrorCode.

This CL lands server change 145087791 by bnc.

BUG=488484

Review-Url: https://codereview.chromium.org/2677493002
Cr-Original-Commit-Position: refs/heads/master@{#447820}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bad0d77fc89ded426e7ee6723fdd3d4433ea5325