Change ErrorCodeToString() argument type from uint32 to ErrorCode.
Not only does this make sense semantically, but this also allows the compiler to
catch if an enum value is not handled in the switch() (not that I would expect
this enum to gain new values).

This CL lands server change 145536832 by bnc.

BUG=488484

Review-Url: https://codereview.chromium.org/2676473005
Cr-Original-Commit-Position: refs/heads/master@{#448137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fadbd7c7b46dc4b338dbe3e2b1bb437a208bba75