Fix Huffman tree construction.
HuffmanBuilder::ToVector() requires that there are at least 2 values
in the tree to create the array representation. This CL updates
HuffmanBuilder::BuildTree() to ensure it always returns a tree containing
at least 2 values.

Also demotes 2 CHECKs to DCHECKs because failure to align with the style guide
as failure does not constitute a security issue and enforces the requirement that
characters should not be larger than 127.

BUG=595493

Review-Url: https://codereview.chromium.org/2694363003
Cr-Original-Commit-Position: refs/heads/master@{#450910}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af126ad510eb5f6ba94fa96ef44bb49a576eaff6